pdf_service_flask/requirements.txt
Flask==2.3.3
Jinja2==3.1.6
weasyprint==65.0
qrcode==8.1
pillow==11.1.0
requests==2.31.0
